QLST/script/requirements.txt
mysql-connector-python
PyMySQL
typing-extensions
